+++ server/schema/customer/res.js
@@ -2,7 +2,7 @@ import {Customer} from '../../models'
 import GraphQLDate from 'graphql-date'
 import bcrypt from 'bcrypt'
 import jwt from 'jsonwebtoken'
-import {_auth, _authAdmin} from '../../util'
+import {_auth, _authAdmin, authGiftCard} from '../../util'
 const SALT = 10
 async function generateLoginJwt({username}, msg = '') {
   var user = await Customer.findOne({where: {username}})
@@ -112,6 +112,18 @@ const resolvers = {
       input.password = bcrypt.hashSync(input.password, SALT)
       return await Customer.create(input).then(async function(customer) {
         return customer
+      })
+    },
+    async applyGiftCard(_, {input}, {loggedInUser}) {
+      _auth(loggedInUser)
+      var giftCard = authGiftCard(input.jwt)
+      if (giftCard.expired) {
+        throw new Error('This gift card has expired!')
+      }
+      var user = await Customer.findById(input.customerId)
+      var balance = user.get('balance') + giftCard.giftCard.amount
+      return await user.update({balance}).then(user => {
+        return user.get('balance')
       })
     },
   },